add upload scripts

+++ workflows/10-upload-archive.ts
@@ -14,13 +14,18 @@ export default async function uploadCurrentData(_options: RunOptions) {
   await fs.ensureDir(getArchivePath());
   let total = 0;
   for await (const entry of fs.walk(getArchivePath())) {
-    if (entry.isFile) {
+    if (entry.isFile && entry.path.endsWith(".json")) {
       // Set the parameters for the object to upload.
       const ext = path.extname(entry.path);
       const contentTypeString = contentType(ext);
       await s3Bucket.putObject(entry.path, await Deno.readFile(entry.path), {
         contentType: contentTypeString,
       });
+      if (total % 100 === 0) {
+        log.info(`uploaded ${total} archived files`);
+      }
+      // remove file
+      await Deno.remove(entry.path);
       total++;
     }
   }
